Add files via upload
I believe this to be a more efficient code all around. It uses matrixes and arrays much more primarily in the calibration. improved long-range if statement. Maybe a piece of timing code to prevent fan mode, although it is variable, hopefully, 4 seconds will cover its variability, but this will need to be tested.